requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pillow
numpy
python-multipart
